backend/requirements.txt
fastapi
psycopg2-binary
requests
pydantic